Put PNG decoding in its own module
