+++ src/texture/mod.rs
@@ -1,6 +1,5 @@
-use std::io;
-
 pub mod tga;
+pub mod png;
 
 #[derive(Clone, Debug)]
 pub enum CompressedTexture {
@@ -12,38 +11,8 @@ pub enum CompressedTexture {
 impl CompressedTexture {
     pub fn decode(&self) -> Texture {
         match self {
-            CompressedTexture::Png(data) => {
-                use image::ImageDecoder;
-                let cursor = io::Cursor::new(data);
-                let decoder = image::codecs::png::PngDecoder::new(cursor).unwrap();
-                let (width, height) = decoder.dimensions();
-                let mut data = vec![0_u8; decoder.total_bytes() as usize];
-                let color_type = decoder.color_type();
-                decoder.read_image(&mut data).unwrap();
-                let data = match color_type {
-                    image::ColorType::Rgba8 => data,
-                    image::ColorType::Rgb8 => {
-                        let rgb = image::ImageBuffer::from_raw(width, height, data).unwrap();
-                        let dynamic = image::DynamicImage::ImageRgb8(rgb);
-                        let rgba = dynamic.into_rgba8();
-                        rgba.into_vec()
-                    }
-                    _ => panic!("Unknown color type {color_type:?}"),
-                };
-                Texture::Rgba {
-                    width,
-                    height,
-                    data,
-                }
-            }
-            CompressedTexture::Tga(data) => {
-                let (width, height, data) = tga::decode(data);
-                Texture::Rgba {
-                    width,
-                    height,
-                    data,
-                }
-            }
+            CompressedTexture::Png(png) => png::decode(png),
+            CompressedTexture::Tga(tga) => tga::decode(tga),
             CompressedTexture::Bc7(_) => todo!("BC7 is still unimplemented"),
         }
     }
